Blockreachpart1 (#44)
* Correct the block state if PlayerActionAckPacket sends it

Originally we did not translate if this packet failed. With this commit, a lot (if not all) of the ghost block issues with mobile devices should be solved if the player is running a new enough version that supports this packet. If the packet returns not successful, we reset the block back to its original type.

This commit also includes cleanup in BlockChangeTranslator as well as proper Netherite tool support for calculating block breaking.

* Correct the block state if PlayerActionAckPacket sends it

Originally we did not translate if this packet failed. With this commit, a lot (if not all) of the ghost block issues with mobile devices should be solved if the player is running a new enough version that supports this packet. If the packet returns not successful, we reset the block back to its original type.

This commit also includes cleanup in BlockChangeTranslator as well as proper Netherite tool support for calculating block breaking.

* Client-side block breaking

* Make block breaking entirely Geyser-side

Don't rely on ServerPlayerActionAckPacket; just do it ourselves.

* Remove no-longer-needed Session variable

* More touches

* Allow blocks to be broken after fire is broken if gamemode is not creative

* Hacky fix for block placing

* No cache chunks is still OK; fix continue break particles

* Fix mobile players being able to reach from far away

* Fix block range checks

* Fix block breaking range

* Add Paper-style check

* Fix block breaking range

Co-authored-by: DoctorMacc <toy.fighter1@gmail.com>